refactor: Repository 패키지 구조 개선 (author/receiver 분리) 및 관련 리팩터링
- Domain 및 Data 레이어의 Repository들을 기능 역할에 따라 `author`와 `receiver` 패키지로 분리 및 이동
- `AfternoteRepositoryModule`을 `AfternoteAuthorRepositoryModule`로 이름을 변경하고, `AuthorReceiverRepository` 바인딩을 해당 모듈로 이전
- `AfternoteHomeScreen`에서 `onFabClick` 파라미터를 nullable로 변경하고, 값이 null일 경우 FAB가 노출되지 않도록 UI 로직 수정
- 패키지 경로 변경에 따른 ViewModel, UseCase, DataSource 내 임포트 및 KDoc 참조 업데이트
- `AfternoteAuthoringFailureMapper` 등 관련 유틸리티 클래스 위치 조정 및 정리
