+++ feature/afternote/data/src/main/kotlin/com/afternote/feature/afternote/data/di/AfternoteReceiverRepositoryModule.kt
@@ -1,9 +1,7 @@
 package com.afternote.feature.afternote.data.di
 
-import com.afternote.feature.afternote.data.repositoryimpl.receiver.AuthorReceiverRepositoryImpl
 import com.afternote.feature.afternote.data.repositoryimpl.receiver.ReceiverRepositoryImpl
-import com.afternote.feature.afternote.domain.repository.AuthorReceiverRepository
-import com.afternote.feature.afternote.domain.repository.ReceiverRepository
+import com.afternote.feature.afternote.domain.repository.receiver.ReceiverRepository
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -16,8 +14,4 @@ abstract class AfternoteReceiverRepositoryModule {
     @Binds
     @Singleton
     abstract fun bindReceiverRepository(impl: ReceiverRepositoryImpl): ReceiverRepository
-
-    @Binds
-    @Singleton
-    abstract fun bindAuthorReceiverRepository(impl: AuthorReceiverRepositoryImpl): AuthorReceiverRepository
 }
