+++ feature/afternote/data/src/main/kotlin/com/afternote/feature/afternote/data/repositoryimpl/author/MemorialPhotoUploadRepositoryImpl.kt
@@ -1,7 +1,7 @@
-package com.afternote.feature.afternote.data.repositoryimpl
+package com.afternote.feature.afternote.data.repositoryimpl.author
 
 import com.afternote.core.domain.repository.PhotoUploadRepository
-import com.afternote.feature.afternote.domain.repository.MemorialPhotoUploadRepository
+import com.afternote.feature.afternote.domain.repository.author.MemorialPhotoUploadRepository
 import javax.inject.Inject
 
 private const val DIRECTORY_AFTERNOTES = "afternotes"
