+++ feature/afternote/domain/src/main/java/com/afternote/feature/afternote/domain/repository/author/MemorialPhotoUploadRepository.kt
@@ -1,4 +1,4 @@
-package com.afternote.feature.afternote.domain.repository
+package com.afternote.feature.afternote.domain.repository.author
 
 /**
  * 추모(플레이리스트) 영정 사진을 content URI에서 업로드하고 HTTPS URL을 반환합니다.
